Changes for cleaner compile on 64-bit architectures. Improved support for Apple Mac. compiler: GNU GCC now supports "assert_static" (so this is now used). compiler: @start is an alias for main (unfinished). compiler: check for common error of forgetting "if" in an "else if (expr)" clause. compiler: in error/warning messages, show the "unexpected" value or name, instead of just saying "unexpected number" or "unexpected symbol". compiler: bug fix, special case of function with only a variable argument list. compiler: bug fix, code generation bug for a mismatch of the number of parameters. compiler: bug fix, compound block on a single line caused error 1 (expecting ";") compiler: bug fix, a packed pseudo-array was considered unpacked because the pseudo-array inherits most fields from the parent symbol. compiler: bug fix in test for UTF-8 files. amxcons: corrections in Win32 terminal support. amxfile: Switch to most recent version of minIni. pawndisasm: generate symbolic labels as jump targets. pawndbg: modifications for interfacing with new IDE (wxQuincy). pawnrun: make time-measuring optional.

+++ amx/fpattern.c
@@ -54,10 +54,10 @@
 *       dtribble@flash.net
 *
 * Copyright 1997-1998 by David R. Tribble, all rights reserved.
-* Copyright (c) 2005-2012, ITB CompuPhase, all rights reserved.
+* Copyright (c) 2005-2015, ITB CompuPhase, all rights reserved.
 * www.compuphase.com
 *
-* Version: $Id: fpattern.c 4732 2012-06-21 16:02:52Z thiadmer $
+* Version: $Id: fpattern.c 5181 2015-01-21 09:44:28Z thiadmer $
 */
 
 
@@ -927,7 +927,7 @@ int fpattern_match(const char *pat, const char *fname, int flength, int keepcase
 
     /* Attempt to match pattern against filename */
     if (flength < 0)
-        flength = strlen(fname);
+        flength = (int)strlen(fname);
     if (flength == 0)
         return (pat[0] == '\0');    /* Special case */
     if (keepcase)
@@ -988,7 +988,7 @@ int fpattern_matchn(const char *pat, const char *fname, int flength, int keepcas
 
     /* Attempt to match pattern against filename */
     if (flength < 0)
-        flength = strlen(fname);
+        flength = (int)strlen(fname);
     if (keepcase)
         rc = fpattern_submatch(pat, fname, flength);
     else
@@ -1012,7 +1012,7 @@ int fpattern_matchcount(const char *pat, const char *fname, int flength, int min
 
     /* Attempt to match pattern against filename */
     if (flength < 0)
-        flength = strlen(fname);
+        flength = (int)strlen(fname);
 
     if (keepcase) {
         for (len = flength; len > minlength; len--)
